+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - update @chevre/api-nodejs-client
 - 予約の余分確保分をsubReservationとして表現するように調整
 - デフォルトで確定予約へ連携する値を拡張
+- デフォルトで注文識別子にconfirmationNumberとconfirmationPassを追加
 
 ### Deprecated